+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+### Added
+- Added `errorResults` getter for `TransactionError` to simplify deriving operation error codes and messages
+
 ## [1.4.2] - 2019-04-05
 ### Fixed
 - Add missing `referrerId` arg to `wallets.create()`